requirements.txt
flask
requests
from markupsafe import Markup

# YOLOv5 requirements
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch
ultralytics

torchvision>=0.8.1
tqdm>=4.64.0
seaborn>=0.11.0
pandas

ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation